refactor(modules): unify backend dispatch on public MINIWORLD/PYTORCH + internal KernelBackend
Public module option is now ImplementationType.{PYTORCH, MINIWORLD} (+CUEQUIVARIANCE
baseline); MINIWORLD = "ours (auto)" picks the fastest correct kernel per GPU/shape/mode.
Concrete tech backends (TRITON/CUTE/CUDA) are internal — resolved in the new central
modules/dispatch.py to a KernelBackend and dispatched on that; forward no longer branches
on ImplementationType. They stay accepted only for explicit per-impl benchmarking.

- New modules/dispatch.py: KernelBackend enum, single-source capability()/is_sm100()/
  is_sm90plus(), and per-op resolve_*() encoding the exact prior arch/family policy.
  trimul dispatch.py now delegates here (removes the duplicated capability copy).
- All 10 module classes route forward on a resolved self._backend. Fixes the 4 modules
  that raised on MINIWORLD (AdaptiveLayerNorm, ConditionedTransition, AugmentedAttention,
  BidirectionalTriangleAttention) — MINIWORLD now builds+runs every module.
- Transition correctness-guard on sm_100 (B200): skip the SM90-only hand-CUDA b2b and
  cute paths (which crash on Blackwell), fall to triton; d>=512 (no fused sm_100 kernel
  fits) falls to the correct torch reference instead of an smem-OOM crash. MINIWORLD
  transition now works on B200 for d=128/256/512 without the env crutch.
- int64-harden the non-hot trimul-triton kernels (front.py forward — the flagged
  asymmetry vs its already-hardened backward, gate_elem, tm1/tm2 main) so large-L flat
  offsets don't overflow int32. Hot transition-triton family left int32 (int32-safe at
  L<=1024; hardening it risks perf) and covered by a static guard test.
- tests/: dispatch resolution + arch-mock, import-surface (MINIWORLD builds all), parity
  PYTORCH vs MINIWORLD, int64 static guard. Standalone run_dispatch_checks.py (no pytest dep).

Behavior-preserving: same kernel selected for the same inputs. Verified on B200 — pairformer
trimul_only + full, inference + training, L=256/384/512 all within +-0.4% of baseline
(noise); per-module MINIWORLD-vs-PYTORCH cosine >= 0.99999.

+++ src/miniworld_kernels/kernels/tm1/triton/main.py
@@ -60,7 +60,9 @@ def fused_sigmoid_gate_fwd_kernel(
     row_start = pid_m * BLOCK_M
     col_start = pid_n * BLOCK_N
 
-    offs_m = row_start + tl.arange(0, BLOCK_M)
+    # int64 M-index: offs_m*N (M=B*L*L) overflows int32 at large logical L; weight
+    # offsets (offs_k*N, offs_n) stay int32 (bounded by K/N).
+    offs_m = row_start + tl.arange(0, BLOCK_M).to(tl.int64)
     offs_n = col_start + tl.arange(0, BLOCK_N)
 
     LA_tile = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
@@ -151,7 +153,9 @@ def fused_sigmoid_gate_bwd_kernel(
     row_start = pid_m * BLOCK_M
     col_start = pid_n * BLOCK_N
 
-    offs_m = row_start + tl.arange(0, BLOCK_M)
+    # int64 M-index: offs_m*N (M=B*L*L) overflows int32 at large logical L; weight
+    # offsets (offs_k*N, offs_n) stay int32 (bounded by K/N).
+    offs_m = row_start + tl.arange(0, BLOCK_M).to(tl.int64)
     offs_n = col_start + tl.arange(0, BLOCK_N)
 
     LA_tile = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
